authorize

+++ src/main/java/com/phatdo/blog/resourceserver/services/ReplyService.java
@@ -4,14 +4,20 @@ import com.phatdo.blog.resourceserver.exception.CustomError;
 import com.phatdo.blog.resourceserver.exception.CustomException;
 import com.phatdo.blog.resourceserver.models.replies.Reply;
 import com.phatdo.blog.resourceserver.models.users.User;
+import com.phatdo.blog.resourceserver.models.users.UserRole;
 import com.phatdo.blog.resourceserver.repositories.BlogRepository;
 import com.phatdo.blog.resourceserver.repositories.ReplyRepository;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
 
+/**
+ * Reply service
+ * @author phatdo 29/08/2024
+ */
 @Service
 public class ReplyService {
     private final ReplyRepository replyRepository;
@@ -31,7 +37,8 @@ public class ReplyService {
                     user.getReplies().add(reply);
                     blog.getReplies().add(reply);
                     return replyRepository.save(reply);
-                }).orElseThrow(() -> new CustomException(CustomError.BLOG_NOT_FOUND));
+                })
+                .orElseThrow(() -> new CustomException(CustomError.BLOG_NOT_FOUND));
     }
 
     public List<Reply> findByBlog(long id) throws CustomException {
@@ -39,26 +46,30 @@ public class ReplyService {
                 .map(blog -> replyRepository.findByBlogId(blog.getId())).orElseThrow(() -> new CustomException(CustomError.BLOG_NOT_FOUND));
     }
 
-    @PreAuthorize("hasAuthority('ROLE_ADMIN') || authentication.principal.username == #user.username")
     public Reply updateReply(long id, String content, User user) throws CustomException {
-        return replyRepository.findById(id)
-                .map(reply -> {
+        Optional<Reply> optReply = replyRepository.findById(id);
+        if (optReply.isPresent()) {
+            Reply reply = optReply.get();
+            if (!reply.getUser().getUsername().equals(user.getUsername()))
+                throw new CustomException(CustomError.ACCESS_DENIED);
             reply.setContent(content);
             return replyRepository.save(reply);
-        })
-                .orElseThrow(() -> new CustomException(CustomError.REPLY_NOT_FOUND));
+        }
+        else
+            throw new CustomException(CustomError.REPLY_NOT_FOUND);
     }
 
-    @PreAuthorize("hasAuthority('ROLE_ADMIN') || authentication.principal.username == #user.username")
     public void deleteReply(long id, User user) throws CustomException {
         Optional<Reply> optReply = replyRepository.findById(id);
         if (optReply.isPresent()) {
             Reply reply = optReply.get();
-            reply.getBlog().getReplies().remove(reply);
-            user.getReplies().remove(reply);
-            replyRepository.delete(reply);
+            if (reply.getUser().getUsername().equals(user.getUsername()) || user.getRoles().contains(UserRole.ADMIN)) {
+                reply.getBlog().getReplies().remove(reply);
+                reply.getUser().getReplies().remove(reply);
+                replyRepository.delete(reply);
+            }
+            else throw new CustomException(CustomError.ACCESS_DENIED);
         }
-        else
-            throw new CustomException(CustomError.REPLY_NOT_FOUND);
+        else throw new CustomException(CustomError.REPLY_NOT_FOUND);
     }
 }
